fix: CIのFail修正 - ページコンポーネントの分離
