+++ src/ui/src/app/page.tsx
@@ -1,22 +1,6 @@
-'use client';
-
-import { useEffect } from 'react';
-import { useRouter } from 'next/navigation';
+import { redirect } from 'next/navigation';
 
 export default function HomePage() {
-  const router = useRouter();
-
-  useEffect(() => {
-    // ルートページにアクセスされたら、アップロードページにリダイレクト
-    router.replace('/upload');
-  }, [router]);
-
-  return (
-    <div className="flex items-center justify-center h-full">
-      <div className="text-center">
-        <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-blue-600 mx-auto mb-4"></div>
-        <p className="text-gray-600">リダイレクト中...</p>
-      </div>
-    </div>
-  );
+  // サーバーサイドでのリダイレクト
+  redirect('/upload');
 }
